requirements.txt
mysql-connector-python
flask
flask-cors
flask-restplus
flasgger